pyproject.toml
[project]
name = "py-winget-source"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.15",
    "pyyaml>=6.0",
]
